safari-534.53-branch/LayoutTests/fast/css/string-quote-binary-expected.txt

This test checks if CSS string values are correctly serialized when they contain binary characters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testElement.style.fontFamily is "'\\0\\1\\2\\3\\4\\5\\6\\7\\8\\9\\a\\b\\c\\d\\e\\f'"
PASS testElement.style.fontFamily is "'\\10\\11\\12\\13\\14\\15\\16\\17\\18\\19\\1a\\1b\\1c\\1d\\1e\\1f\\7f'"
PASS testElement.style.fontFamily is "'\\a\\b\\c\\d\\e\\f\\1a\\1b\\1c\\1d\\1e\\1f\\7f'"
PASS testElement.style.fontFamily is "'\\0\\1\\2'"
PASS testElement.style.fontFamily is "'\\0  \\1  \\2  '"
PASS testElement.style.fontFamily is "'\\0   \\1   \\2   '"
PASS testElement.style.fontFamily is "'\\f\\0g'"
PASS testElement.style.fontFamily is "'\\0 0\\0 1\\0 2\\0 3\\0 4\\0 5\\0 6\\0 7\\0 8\\0 9'"
PASS testElement.style.fontFamily is "'\\0 A\\0 B\\0 C\\0 D\\0 E\\0 F\\0G'"
PASS testElement.style.fontFamily is "'\\0 a\\0 b\\0 c\\0 d\\0 e\\0 f\\0g'"
PASS successfullyParsed is true

TEST COMPLETE
